fixed crach if 'HOME' variable is not defined
git-svn-id: https://only.mawhrin.net/repos/stuff/trunk@1731 62a0de91-87f2-0310-aeaa-8eaa5733e38e

+++ zlibrary/src/abstract/filesystem/ZLFile.cpp
@@ -32,8 +32,7 @@
 ZLFile::ZLFile(const std::string &path) : myPath(path), myInfoIsFilled(false) {
 	ZLFSManager::instance().normalize(myPath);
 	{
-		int index = ZLFSManager::instance().findLastFileNameDelimiter(myPath);
-		// TODO: remove this hack
+		size_t index = ZLFSManager::instance().findLastFileNameDelimiter(myPath);
 		if (index < myPath.length() - 1) {
 			myNameWithExtension = myPath.substr(index + 1);
 		} else {
